Move miniCUTE parsers into a submodule
