+++ app/Main.hs
@@ -3,7 +3,7 @@ module Main
   ) where
 
 import Data.Text.Prettyprint.Doc.Minicute ( prettyShow )
-import Minicute.Parser.Parser ( mainProgramL )
+import Minicute.Parser.Minicute.Parser ( mainProgramL )
 import Minicute.Transpilers.LambdaLifting ( lambdaLifting )
 import System.Environment
 import System.IO
